+++ election-server/resources/schema.sql
@@ -11,12 +11,12 @@ CREATE TABLE IF NOT EXISTS ballots (
   id INT8 PRIMARY KEY, -- For pagination
   transaction_hash BYTEA NOT NULL,
   block_time TIMESTAMP WITH TIME ZONE NOT NULL,
-  ballot JSONB NOT NULL,
+  ballot BYTEA NOT NULL,
   account BYTEA NOT NULL,
   verified BOOL NOT NULL
 );
 
--- Table containing voting power delegations
+-- Table containing voting weight delegations
 CREATE TABLE IF NOT EXISTS delegations (
   id INT8 PRIMARY KEY, -- For pagination
   transaction_hash BYTEA NOT NULL,