+++ coordinator-tool/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.2.4
 
 - Parallelize the tally computation.
+- Updates concordium rust dependencies
 
 ## 0.2.3